organized code for multi page dash and made buttons 2 px smaller
