+++ assets/style.css
@@ -62,7 +62,7 @@ h1 {
     #light-btn,
     #feed-btn {
         
-        height: 53px;
+        height: 48px;
 
     }
 }
